+++ tests/system_tests/test_launch.py
@@ -224,3 +224,49 @@ class LifecycleTests(LifecycleTestsBase):
                 session.send_request('disconnect')
 
             adapter.wait()
+
+    def test_launch_ptvsd_client_with_output(self):
+        argv = []
+        waitscript = dedent("""
+            import sys
+            sys.stdout.write('ok')
+            sys.stderr.write('ex')
+            """)
+        options = {'debugOptions':['RedirectOutput']}
+        filename = self.write_script('spam.py', waitscript)
+        script = self.write_debugger_script(filename, 9876, run_as='script')
+        with DebugClient(port=9876) as editor:
+            adapter, session = editor.host_local_debugger(
+                argv,
+                script,
+
+            )
+            with session.wait_for_event('exited'):
+                with session.wait_for_event('thread'):
+                    (req_initialize, req_launch, req_config, _, _, _
+                     ) = lifecycle_handshake(session, 'launch', options=options)
+
+                adapter.wait()
+
+        # Skipping the 'thread exited' and 'terminated' messages which
+        # may appear randomly in the received list.
+        received = list(_strip_newline_output_events(session.received))
+        self.assert_received(received, [
+            self.new_version_event(session.received),
+            self.new_response(req_initialize, **INITIALIZE_RESPONSE),
+            self.new_event('initialized'),
+            self.new_response(req_launch),
+            self.new_response(req_config),
+            self.new_event('process', **{
+                'isLocalProcess': True,
+                'systemProcessId': adapter.pid,
+                'startMethod': 'launch',
+                'name': filename,
+            }),
+            self.new_event('thread', reason='started', threadId=1),
+            self.new_event('output', category='stdout',  output='ok'),
+            self.new_event('output', category='stderr',  output='ex'),
+            self.new_event('thread', reason='exited', threadId=1),
+            self.new_event('exited', exitCode=0),
+            self.new_event('terminated'),
+        ])
